requirements.txt
asgiref==3.8.1
coverage==7.8.0
Django==5.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.8.0
pytest==7.4.3
pytest-cov==6.1.0
pytest-django==4.7.0
pytz==2025.2
setuptools==78.1.0
sqlparse==0.5.3
